Brand new testing tool
An updated version of the keyboard.py tool that allows to change the message type without the need to change the code. I'm keeping the old code because this is not yet compatible with all messages from std_msgs (manily missing multi dimensional arrays) nor custom messages and the old code is easier to change without breaking it. However, once i figure out how to include arrays and custom messages i'll update the old file with this one.